refactor: 회원가입 데이터 변환 로직을 컴포넌트에서 useSignup 훅으로 이동하여 SRP 적용
